Changing personal info.
After debating with my father (@ebalter) about using "an" or "a" in "I am [a|an] human", he convinced me I was wrong, so I had to use "a" instead of "an".

One thing I don't understand is why "hour" begins with a vowel sound and "human" don't.

If you want to look further on this, these links bellow are good references:
- http://www.aoran.org/
- http://www.aoran.org/a-or-an-before-h.html
- http://www.thefreedictionary.com/human